requirements.txt
gradio
bs4
langchain
langchain_community
langchain_google_genai
langchain-huggingface
langchain_text_splitters
python-dotenv
protobuf==3.20.*
setuptools
faiss-cpu
sentence-transformers